refactor: use ros message types in radio communication
